+++ UI/src/views/AdminViews/SubjectsView.vue
@@ -1,0 +1,3 @@
+<template></template>
+<script setup></script>
+<style  scoped></style>
